@@ -1,6 +1,4 @@
-﻿using Newtonsoft.Json;
-
-#nullable disable
+﻿using System.Text.Json.Serialization;
 
 namespace LabWebApi.Repository.Entities
 {
