@@ -1,7 +1,5 @@
 ﻿using System.Collections.Generic;
-using Newtonsoft.Json;
-
-#nullable disable
+using System.Text.Json.Serialization;
 
 namespace LabWebApi.Repository.Entities
 {
